% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Looks up the most recent USD/EUR exchange rate and compute the provided USD amount in EUR (as a number)}
\usage{
convert_usd_to_eur(usd, decimals = 3)
}
\arguments{
\item{decimals}{number of decimals for the exchange rate}

\item{x}{number of USD}
}
\value{
integer
}
\description{
Looks up the most recent USD/EUR exchange rate and compute the provided USD amount in EUR (as a number)
}
\examples{
convert_usd_to_eur(10)
convert_usd_to_eur(100)
}
