% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertcurr.R
\docType{package}
\name{convertcurr}
\alias{convertcurr}
\alias{convertcurr-package}
\title{convertcurr package}
\description{
Contains five functions related to EUR conversion
}
\author{
Anna Alexandra Grigoryan \email{anna.alexandra.grigoryan@gmail.com}
}
