% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\name{eur}
\alias{eur}
\title{Formats a number to a string using the Euro sign, rounding up to 2 digits and using the comma as the big.mark}
\usage{
eur(amount)
}
\arguments{
\item{amount}{amount to transform to €}
}
\value{
string
}
\description{
Formats a number to a string using the Euro sign, rounding up to 2 digits and using the comma as the big.mark
}
\examples{
eur(675687)
eur(-8787876)
}
\seealso{
uneur()
}
