% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{Looks up the most recent USD/EUR exchange rate via an API call and returns the rate as a number.}
\usage{
eurusd(decimals = 3)
}
\arguments{
\item{decimals}{number of decimals to show}
}
\value{
integer
}
\description{
Looks up the most recent USD/EUR exchange rate via an API call and returns the rate as a number.
}
\examples{
eurusd()
eurusd(2)
}
