% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{Takes two arguments (date_from and date_to) and returns a data.table object on the daily rates from date_from to the date_to dates provided with 2 columns (date and exchange rate)}
\usage{
eurusds(date_from, date_to, decimals = 3)
}
\arguments{
\item{date_from}{starting date of the period, back-filled for missing dates}

\item{date_to}{ending date of the period, back-filled for missing dates}

\item{decimals}{number of decimals to show}
}
\value{
integer
}
\description{
Takes two arguments (date_from and date_to) and returns a data.table object on the daily rates from date_from to the date_to dates provided with 2 columns (date and exchange rate)
}
\examples{
eurusds('2019-01-01', '2019-01-10')
eurusds('2018-02-01', '2018-02-10')
}
