% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{Formats a string  to a number.}
\usage{
uneur(amount)
}
\arguments{
\item{amount}{amount to format to number}
}
\value{
string
}
\description{
Formats a string  to a number.
}
\examples{
eur("€6,565,757")
eur("-€177,343")
}
\seealso{
uneur()
}
